Add munged "--trusted-group" cmdline opt
MUNGE directories must be owned by the same user as the munged daemon
process.  They cannot allow write permissions for group or other unless
the sticky bit is set.  In addition, all of the parent directories in
the path on up to the root directory must be owned by either root or
the same user as the munged daemon process.  They also cannot allow
write permissions for group or other unless the sticky bit is set.
This prevents unprivileged users from removing or renaming a file in
the trusted directory, or replacing a file or directory in the path
leading to a trusted directory.

This creates a problem if MUNGE is installed under a filesystem
hierarchy owned by an administrative group (e.g., /opt or /usr/local).
If this directory hierarchy allows write permissions for group
without setting the sticky bit, munged will refuse to start unless
the "--force" command-line option is given.  But this option is not
intended for use in production since it overrides other warnings that
can affect security.

This commit adds the "--trusted-group <group_name_or_gid>" command-line
option to munged.  Directories with group write permissions are allowed
if they are owned by the trusted group (or the sticky bit is set).

Tested on:
- CentOS 6.9, 7.3.1611
- Debian 9.0
- Fedora 25
- FreeBSD 11.0
- NetBSD 7.1
- OpenBSD 4.7, 6.1
- Ubuntu 16.04.2 LTS, 17.04

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>