Switch query_uid to use xgetpwnam
Back in dfe5ea5, occasional SEGVs in munged were traced back to the use
of getgrent() & getpwnam() in a routine where these non-reentrant calls
should have been safe to use.  Switching to the reentrant versions
addressed these SEGVs but necessitated the use of a portability
abstraction.

This commit replaces the use of getpwnam() within query_uid() with
the xgetpwnam() abstraction in munged.  This abstraction is used by
munged directly, and now by munge & remunge via query_uid().

In xgetpwnam.c, fallback code exists in case no suitable version of
getpwnam_r() exists on the system.  This conditionally-compiled code
exists in blocks surrounded by the HAVE_GETPWNAM preprocessor macro.
It locks a static mutex, calls getpwnam(), copies the data from
getpwnam()'s static buffer, and unlocks the mutex.  It presumes it
is being linked against pthreads; however, the munge executable is
not threaded.  Since this code must now be built with pthreads (for
munged and remunge) and without pthreads (for munge), this commit
conditionally compiles these mutex calls based on the WITH_PTHREADS
preprocessor macro.

Since xgetpwnam.c & xgetpwnam.h are now shared, it makes sense to
move them to libcommon.  But since they must now be compiled both
with and without support for pthreads, they are specified in the
automake sources (munge_SOURCES, remunge_SOURCES, and munged_SOURCES)
with relative paths.  Similarly, query.c and query.h have been moved
out of libcommon (but still reside in that directory) and are listed
in the automake sources using relative paths.  These files will be
reorganized in a later commit; they are not being moved now since
that would obscure the modifications made to routines within.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>