Explicitly init getpwnam_r & getgrent_r result ptrs
The xgetpwnam() function is a portable encapsulation of getpwnam_r().
The last parameter of the POSIX variant of getpwnam_r() is a result
pointer.  It stores a pointer to the passwd structure if a matching
name is found, or is set to NULL if no match is found or an error
occurs.  The getgrnam_r() function is defined similarly.

While working on xgetgrnam(), it was discovered that getgrnam_r()
on OpenBSD 6.1 does not explicitly set this result pointer to NULL
on ERANGE (although it does correctly set it to NULL if no match
is found).  While OpenBSD 6.1 does indeed set the getpwnam_r()
result pointer to NULL on ERANGE, this commit explicitly sets it to
NULL before calling the POSIX variant of getpwnam_r() as a purely
defensive measure.

Likewise, this commit also sets the result pointer to NULL before
calling the GNU variant of getgrent_r() .

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>